Bugfix: Fixed NSTextAlignment rawValue conversion to Int

+++ BXUIKit/NS-UI Compatibility/NSUITextView.swift
@@ -249,7 +249,7 @@ extension NSTextAlignment
 	{
 		var value = UInt(intValue)
 		value = NSTextAlignment.swapRightAndCenter(value)
-		self.init(rawValue:value)
+        self.init(rawValue:Int(value))
 	}
 	
 	/// The order of intValue follows iOS conventions, i.e. left, center, right, justified.
@@ -257,7 +257,7 @@ extension NSTextAlignment
 	public var intValue:Int
 	{
 		var value = self.rawValue
-		value = NSTextAlignment.swapRightAndCenter(value)
+        value = Int(NSTextAlignment.swapRightAndCenter(UInt(value)))
 		return Int(value)
 	}
 	
@@ -265,11 +265,11 @@ extension NSTextAlignment
 	{
 		if value == NSTextAlignment.right.rawValue
 		{
-			return NSTextAlignment.center.rawValue
+            return UInt(NSTextAlignment.center.rawValue)
 		}
 		else if value == NSTextAlignment.center.rawValue
 		{
-			return NSTextAlignment.center.rawValue
+            return UInt(NSTextAlignment.center.rawValue)
 		}
 		
 		return value
